feat(settings): there is a default texture importer and user can change it
also move the Settings script in a folder
