+++ package/Editor/Presets.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0d04b991eaf17a643b5aa1baf51ac077
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
