+++ package/Editor/Presets/Scenario Default TextureImporter.preset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 28269680c775243409a2d470907383f9
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2655988077585873504
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
