+++ package/Editor/Settings.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 08d4aab7cf61e054fba2b37e5c318c83
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
